Added TicketsListData; js_only.
