+++ assets/prototype/domain/eventEditor/tickets/ticketsList/TicketsListData.tsx
@@ -1,0 +1,33 @@
+/**
+ * Internal dependencies
+ */
+import AddNewTicketButton from './AddNewTicketButton';
+import EmptyState from '../../../../application/ui/components/display/EmptyState';
+import ErrorIndicator from '../../../../application/ui/components/display/ErrorIndicator';
+import LoadingIndicator from '../../../../application/ui/components/display/LoadingIndicator';
+import TicketsList from './TicketsList';
+import useTickets from '../../data/queries/tickets/useTickets';
+import { useStatus, TypeName } from '../../../../application/services/apollo/status';
+
+const TicketsListData = () => {
+	const tickets = useTickets();
+	const { isError, isLoading } = useStatus();
+	const error = isError(TypeName.tickets);
+	const loading = isLoading(TypeName.datetimes) || isLoading(TypeName.tickets);
+
+	if (loading) return <LoadingIndicator message='loading tickets...' />;
+
+	if (error) return <ErrorIndicator error={error} />;
+
+	if (!tickets.length) {
+		return (
+			<EmptyState description={'try changing filter settings'} title={'NO TICKETS FOR YOU !!!'}>
+				<AddNewTicketButton />
+			</EmptyState>
+		);
+	}
+
+	return <TicketsList tickets={tickets} />;
+};
+
+export default TicketsListData;
