+++ assets/prototype/domain/eventEditor/tickets/ticketsList/filterBar/TicketListFilterBar.js
