+++ assets/prototype/domain/eventEditor/tickets/ticketsList/filterBar/controls/DisplayTicketsControl.js
