+++ assets/prototype/domain/eventEditor/tickets/ticketsList/filterBar/controls/TicketsChainedButton.js
