+++ assets/prototype/domain/eventEditor/tickets/ticketsList/filterBar/controls/TicketsSortedByControl.js
